+++ src/backend/ci/core/common/common-api/src/main/kotlin/com/tencent/devops/common/api/expression/GrammarAnalysis.kt
@@ -25,7 +25,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.process.engine.control.expression
+package com.tencent.devops.common.api.expression
 
 import java.lang.Exception
 import java.util.ArrayDeque
