+++ src/backend/ci/core/common/common-api/src/main/kotlin/com/tencent/devops/common/api/expression/Word.kt
@@ -25,8 +25,13 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.process.engine.control.expression
+package com.tencent.devops.common.api.expression
 
-import java.lang.Exception
-
-class GrammarAnalysisFinishedException(override val message: String) : Exception(message)
+data class Word(
+    val str : String,
+    val symbol : String
+) {
+    override fun toString(): String {
+        return "Word(str='$str', symbol='$symbol')"
+    }
+}
