+++ src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/engine/control/ControlUtils.kt
@@ -35,7 +35,7 @@ import com.tencent.devops.common.pipeline.enums.JobRunCondition
 import com.tencent.devops.common.pipeline.enums.StageRunCondition
 import com.tencent.devops.common.pipeline.pojo.element.ElementAdditionalOptions
 import com.tencent.devops.common.pipeline.pojo.element.RunCondition
-import com.tencent.devops.process.engine.control.expression.EvalExpress
+import com.tencent.devops.common.api.expression.EvalExpress
 import com.tencent.devops.process.utils.TASK_FAIL_RETRY_MAX_COUNT
 import com.tencent.devops.process.utils.TASK_FAIL_RETRY_MIN_COUNT
 import org.slf4j.LoggerFactory
